sixth commit: added quick sort and merge sort
